requirements.txt
Django==3.0.7
coverage==4.5.4
dj-database-url==0.5.0
django-allauth==0.40.0
django-braces==1.14.0
django-classy-tags==1.0.0  # for cookie-law
django-cookie-law==2.0.3
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.8
django-migration-testcase==0.0.15
django-paypal==1.0.0
#djangorestframework==3.10.3
#dj-static==0.0.6
psycopg2==2.8.4
python-dateutil==2.8.0
#PyYAML==5.1.2
model-bakery==1.0.1
django-ckeditor==5.9.0
pillow==7.1.2
gunicorn==19.7.1
#static3==0.7.0
shortuuid==0.5.0
beautifulsoup4==4.8.1
python-memcached==1.59
#pytest==5.3.0
#pytest-django==3.7.0
#pytest-cov==2.8.1
